+++ generators/app/questions/application.js
@@ -7,11 +7,6 @@ module.exports = [{
   default: 'sails-rest-api'
 }, {
   type: 'input',
-  name: 'application:api-secret-key',
-  message: 'Type your private API key',
-  default: crypto.randomBytes(32).toString('hex')
-}, {
-  type: 'input',
   name: 'application:jwt-secret-key',
   message: 'Type your private key for JSON Web Token',
   default: crypto.randomBytes(32).toString('hex')